+++ src/components/MedicalResult/MedicalResultFormConsulting.jsx
@@ -1,7 +1,8 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import {
   Form,
   Input,
+  Select,
   Button,
   Card,
   Row,
@@ -18,8 +19,10 @@ import {
   ReloadOutlined,
 } from "@ant-design/icons";
 import { submitConsultationResult } from "../../api/medicalResultAPI";
+import api from "../../configs/api";
 
 const { TextArea } = Input;
+const { Option } = Select;
 const { Text } = Typography;
 
 /**
@@ -34,6 +37,32 @@ const MedicalResultFormConsulting = ({
 }) => {
   const [form] = Form.useForm();
   const [loading, setLoading] = useState(false);
+  const [loadingProtocols, setLoadingProtocols] = useState(false);
+  const [treatmentProtocols, setTreatmentProtocols] = useState([]);
+
+  // Fetch treatment protocols
+  const fetchTreatmentProtocols = async () => {
+    try {
+      setLoadingProtocols(true);
+      const response = await api.get('/treatment');
+      setTreatmentProtocols(response.data || []);
+    } catch (error) {
+      console.error("Error fetching treatment protocols:", error);
+      setTreatmentProtocols([]);
+    } finally {
+      setLoadingProtocols(false);
+    }
+  };
+
+  // Load treatment protocols on component mount
+  useEffect(() => {
+    fetchTreatmentProtocols();
+  }, []);
+
+  // Handle field changes
+  const handleFieldChange = (field, value) => {
+    form.setFieldValue(field, value);
+  };
 
   // Dữ liệu mẫu cho form khám bệnh/tư vấn
   const defaultFormData = {
@@ -42,6 +71,7 @@ const MedicalResultFormConsulting = ({
       "Bệnh nhân có triệu chứng ngứa, đau rát vùng kín, có dịch tiết bất thường",
     diagnosis: "Viêm âm đạo do nấm Candida",
     treatmentPlan: "Sử dụng thuốc kháng nấm, tái khám sau 1 tuần",
+    treatmentProtocolId: null,
   };
 
   // Set initial form values
@@ -59,6 +89,7 @@ const MedicalResultFormConsulting = ({
         ...values,
         appointmentDetailId: appointmentDetail?.id || 123,
         resultType: "CONSULTATION",
+        treatmentProtocolId: values.treatmentProtocolId || null,
       };
 
       console.log("[DEBUG] Final submit data:", submitData);
@@ -202,6 +233,34 @@ const MedicalResultFormConsulting = ({
                   rows={6}
                   placeholder="Ví dụ: Sử dụng thuốc kháng nấm, tái khám sau 1 tuần"
                 />
+              </Form.Item>
+            </Col>
+          </Row>
+
+          <Row gutter={24}>
+            <Col span={12}>
+              <Form.Item
+                name="treatmentProtocolId"
+                label="Phác đồ điều trị"
+                extra="Chọn phác đồ điều trị có sẵn (không bắt buộc)"
+              >
+                <Select
+                  placeholder="Chọn phác đồ điều trị..."
+                  loading={loadingProtocols}
+                  allowClear
+                  showSearch
+                  optionFilterProp="children"
+                  filterOption={(input, option) =>
+                    option?.children?.toLowerCase().includes(input.toLowerCase())
+                  }
+                  onChange={(value) => handleFieldChange("treatmentProtocolId", value)}
+                >
+                  {treatmentProtocols.map((protocol) => (
+                    <Option key={protocol.id} value={protocol.id}>
+                      {protocol.diseaseName}
+                    </Option>
+                  ))}
+                </Select>
               </Form.Item>
             </Col>
           </Row>
